Refactor project into hexagonal architecture
