+++ RadioQ10/Domain/Entities/AddSongRequest.cs
@@ -1,6 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace RadioQ10.Models;
+namespace RadioQ10.Domain.Entities;
 
 public sealed class AddSongRequest
 {
